requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
mlflow >=1.20 
optuna
hydra-core >=1.1 
prettytable